third_party/blink/web_tests/http/tests/loading/trust-tokens/trust-token-xhr.tentative.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Tests the prototype Trust Token API's functionality through its XHR interface.</title>
<link rel="help" href="https://github.com/WICG/trust-token-api#trust-token-redemption" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
  'use strict';

  async_test((t) => {
    let request = new XMLHttpRequest();
    request.open('GET', 'https://trusttoken.test');
    request.setTrustToken({
      type: 'token-request'
    });
    request.onerror = t.step_func_done(() => {
      assert_equals(request.trustTokenOperationError.name, "InvalidStateError");
    });
    request.send();
  }, 'No keys for the issuer -> expect issuance to fail');

  async_test((t) => {
    let request = new XMLHttpRequest();
    request.open('GET', 'https://trusttoken.test');
    request.setTrustToken({
      type: 'srr-token-redemption'
    });
    request.onerror = t.step_func_done(() => {
      assert_equals(request.trustTokenOperationError.name, "InvalidStateError");
    });
    request.send();
  }, 'No keys for the issuer -> expect redemption to fail');

  async_test((t) => {
    let request = new XMLHttpRequest();
    request.open('GET', 'https://destination.test');
    request.setTrustToken({
      type: 'send-srr',
      issuer: 'https://issuer.test'
    });
    request.onerror = t.step_func_done(() => {
      assert_equals(request.trustTokenOperationError.name, "OperationError");
    });
    request.send();
  }, 'No SRR for the issuer -> expect signing to fail');
</script>
